chore(test): move e2e test to examples
